+++ R/stats.R
@@ -24,7 +24,7 @@
 #' tok <- tokens_remove(tok, stopwords("en"))
 #' dtm <- dfm(tok, tolower = TRUE)
 #' dtm <- dfm_trim(dtm, min_docfreq = 3)
-#' res <- rainette(dtm, k = 3, min_uc_size = 15)
+#' res <- rainette(dtm, k = 3, min_segment_size = 15)
 #' groups <- cutree_rainette(res, k = 3)
 #' rainette_stats(groups, dtm)
 #' }